Close STDERR when running as a daemon.
When testing on the CentOS8 Stream OS and trying to run the pgaudit_analyze binary as a daemon over an SSH connection, the entire SSH connection would hang and never return. After a lot of investigation, we were able to determine that this was happening when using openssh 8.0p1-9.el8 and also 8.0p1-8.el8. However, when you downgrade openssh to 8.0p1-7.el8, launching the daemon works as it has in the past. After digging through the openssh commits we found this:

On May 2, 2021, CentOS 8 Stream packagers introduced a change to its openssh libraries, which you can see here: https://git.centos.org/rpms/openssh/c/1c3003998fcc4d40fe9e4544bca70dbc03324e65?branch=c8s

After a lot more digging and testing, the issue was related to the fact that STDERR was never closed so in the newer openssh versions it would not release the connection and would hang. After this change we are able to run the pgaudit_analyze daemon over an SSH connection like before.